i386: fix possible stack corruption
The EAX_EDX_RET uses "=A" as result type which implies that the
64-bit variant of the output data is used. On i386, however,
unsigned long is just 32 bit. Thus, the variable holding the time
data may be written beyond the 32 bit limit thus potentially
causing a stack corruption and a crash.

+++ jitterentropy-base-user.h
@@ -99,7 +99,7 @@
 # define EAX_EDX_VAL(val, low, high)     ((low) | (high) << 32)
 # define EAX_EDX_RET(val, low, high)     "=a" (low), "=d" (high)
 #elif __i386__
-# define DECLARE_ARGS(val, low, high)    unsigned long val
+# define DECLARE_ARGS(val, low, high)    unsigned long long val
 # define EAX_EDX_VAL(val, low, high)     val
 # define EAX_EDX_RET(val, low, high)     "=A" (val)
 #endif
